rename: api 호출 훅 위치 변경(#72)
