+++ src/pages/player/CoachPage.tsx
@@ -1,5 +1,5 @@
-import { useCoachList } from '@/assets/hooks/useCoachList';
 import Breadcrumb from '@/features/common/Breadcrumb';
+import { useCoachList } from '@/features/game/hooks/useCoachList';
 import SearchBar from '@/features/media/common/SearchBar';
 import { PlayerList } from '@/features/player/components/';
 
@@ -14,10 +14,14 @@ const CoachPage = () => {
     return <div>{error}</div>;
   }
 
+  const handleSubmit = () => {
+    console.log('TODO: 이벤트 구현');
+  };
+
   return (
     <div>
       <Breadcrumb
-        leftComponent={<SearchBar />}
+        leftComponent={<SearchBar onSubmit={handleSubmit} />}
         paths={[
           { key: 'home', label: 'Home' },
           { key: 'player', label: 'Player' },
