+++ src/pages/player/id/CoachDetailPage.tsx
@@ -1,5 +1,5 @@
-import useCoach from '@/assets/hooks/useCoach';
 import Breadcrumb from '@/features/common/Breadcrumb';
+import useCoach from '@/features/game/hooks/useCoach';
 import { useSearchParams } from 'react-router';
 
 const CoachDetailPage = () => {
